move scripts
