+++ scripts/aip-schema-validation.js
@@ -1,11 +1,11 @@
 const Ajv = require("ajv")
-const validationSchema = require("./aip-schema.json")
-const rawJsonAips = require("./content/ipfs-aips/all-aips.json")
+const validationSchema = require("../aip-schema.json")
+const rawJsonAips = require("../content/ipfs-aips/all-aips.json")
 
 const ajv = new Ajv()
 const validate = ajv.compile(validationSchema)
 
-jsonAips = Object.values(rawJsonAips)
+const jsonAips = Object.values(rawJsonAips)
 for (const jsonAip of jsonAips) {
   delete Object.assign(jsonAip, { description: jsonAip["content"] })["content"]
   const valid = validate(jsonAip)
